requirements.txt
streamlit>=1.32.0
crewai==0.11.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
openai>=1.0.0
pysqlite3-binary>=0.5.0 